drivers/testing/heapcheck: Replace prandom_u32 with custom implementation
prandom_u32 was dropped from 6.1, but get_random_u32 is not oob-safe. So
we only use it for initializing a seed and then rely on our own version
of next_pseudo_random32. The latter is still available via
linux/prandom.h, but given that there are no more in-tree users, it
could be dropped at any point in the future.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>