ci: Switch to native trigger
Now that we are running both pipelines on the same server, we can also
make use of native gitlab-ci trigger mechanism. This both obsoletes the
need to configure a token and avoids starting a real job on a real
runner, just for calling curl.

This also improves the visualization of the two related pipelines in the
gitlab UI.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>