lib/copperplate: Release main threadobj on exit
In case of CONFIG_XENO_PSHARED, the main thread's tcb is allocated on
the shared heap. Normally, threadobjs are released via the destructor
of threadobj_tskey. But this is not called when the main thread
terminates so that we leak the object. Fix that by installing an atexit
handler that only needs to perform the release now.

Reported-by: Mauro S. <mau.salvi@tin.it>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>