y2038: Prepare code base for 64bit time_t
Compiling the Xenomai code base with CFLAGS set to
"-D_TIME_BITS=64 -D_FILE_OFFSET_BITS=64" - which is the requirement for
y2038 safety on 32 bit platforms - failed. We had some problems with
printf format specifiers which are now too small when time_t is lifted
to 64 bit.

Casting to (long long) should do it on all platforms we support.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>