cobalt/posix/select: Fix timeout update in case of -EINTR
When the select() syscall was interrupted (e.g. for handling a Linux
signal) the supplied timeout has not been updated. __cobalt_select()
bailed out too early.

This is not a critical thing as the manpage states:
  On  Linux,  select()  modifies timeout to reflect the amount of time
  not slept; most other implementations do not do this.  (POSIX.1
  permits either behavior.)

  This causes problems both when Linux code
  which reads timeout is ported to other operating systems, and when
  code is ported to Linux that reuses a struct timeval for multiple
  select()s in a loop without reinitializing it.  Consider timeout to
  be undefined after select() returns."

The Xenomai select() implementation is now aligned with Linux.

The testsuite has been improved to cover this case as well.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>