Make uncapped kernel SUBLEVEL available to wrappers
As the SUBLEVEL in LINUX_VERSION_CODE is 8-bits only and is clamped at
255, we cannot detect build-breaking changes that come in later.
Upstream does not care about out-of-tree, so work around this limitation
by extracting the true SUBLEVEL in prepare-kernel.sh and exporting it
via Kconfig. This may only be relevant for 4.19 and 5.10 because 5.12
has a KERNEL_VERSION_SUBLEVEL define in the generated version.h.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>