testsuite: vxworkstests: Do not try to call nonexistent binary
vxworkstests are part of the smokey-testsuite. There is no vxworkstests
executable that can be called.

The following line could be found in the test log:
    sh: 1: /usr/lib/xenomai/testsuite/vxworkstests: not found

This patch addresses the first part of the problem. The remaining part,
the missing error reporting (the testrun was still successful), will be
addressed in a follow-up patch.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>