testsuite: alchemytests: Skip pipe test if kernel support is missing
The pipe test depends on the RTIPC/XDDP protocol which must be provided
by the kernel. Skip the test if the kernel does not support it.

Might be that the module is not loaded or the kernel configuration has
CONFIG_XENO_DRIVERS_RTIPC_XDDP set to 'n'.

The error message reported by smokey:
0"000.543| WARNING: [alchemytests_pi] RTIPC/XDDP protocol not supported by kernel
0"000.567| BUG in __traceobj_assert_failed(): [alchemytests_pi] trace assertion failed:
           pipe-1.c:88 => "ret >= 0"

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>