testsuite: alchemytests: Try loading the xeno_rtipc module
The pipe test depends on the xeno_rtipc kernel module with XDDP
support enabled. We can not be sure that XDDP support is really
available after loading the module, but loading the module is a good
first step.

The test will later be skipped if XDDP support is still not available.

As the pipe test itself is running outside smokey we do not have the
infrastructure around smokey_modprobe() and smokey_rmmod() available.
We have to load/unload the module from smokey.

While at it: Cleaning up some unused includes.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>