compat support: cobalt/arm: Refactor registers for arm alignment
So far, the syscall number for arm is loaded in ARM_ORIG_r0 and ARM_r7
contains the syscall marker XENO_ARM_SYSCALL.
An alignment change is now made with respect to the arm EABI convention
such that ARM_r7 holds the syscall number and therefore r0 to r6 can be
used normally as arguments.
The now unused XENO_ARM_SYSCALL is removed.

With the arm alignment, the specific arm implementation
pipeline_get_syscall_args() for fetching the syscall arguments is not
needed anymore and is replaced by the generic function
syscall_get_arguments().

Signed-off-by: Clara Kowalsky <clara.kowalsky@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>