testsuite: smokey - new testcase pi_nesting_contend_multicore
The testcase ensures that the correct handling of priority boosts in a nested
priority-inheritance scenario in a multi-core environment by assigning different
threads to different cores if this is possible. On single-core targets all
threads are assigned to the same core. In contrast to protect_nesting_protect
and protect_nesting_pi only priority-inheritance is involved here.

For more details please refer to the detailed comment before
pi_nesting_contend_multicore.

Signed-off-by: Fabian Scheler <Fabian.Scheler@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>