cobalt/synch: add [__cobalt_mutex_]lookup_lazy_pp prototype to header
There is kind of a layering violation here between synch and
cobalt/mutex by the latter providing a helper to the former for priority
protection/ceiling. With kernel 6.8 getting picky about missing
prototypes for non-static functions, we need to resolve at least that
aspect. Therefore, establish a prototype in synch.h as it is not easily
possibly to include posix/mutex.h.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>