testsuite/gpiobench: Fix output pin in react mode
The gpiobench was tested with one device in loopback mode and another
device in react mode. The former sends an asynchronous event to the
react device by toggling its output pin. In this test state, however,
the react device always responds with a high signal on its output pin.
Yet the input pin of the loopback device is set to edge-triggered, which
is why the test hangs in read() on both sides after one loop.

For the test to work, the output pin of the react device must toggle
between high and low in order for the loopback device to successfully
read a change on the input pin.

Signed-off-by: Clara Kowalsky <clara.kowalsky@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>