testsuite/gpiobench: Synchronize index of loopback mode and react mode
While the index of the react device is incremented on every round (read,
write, increment), the index of the loopback is only increased on every
2nd round (write, read, sleep, repeat last steps, increment). This
results in the unintuitive behavior that half the number of loops would
have to be passed as a parameter for the loopback device in order for
both devices to stop at the same time.

To simplify this, the index of the loopback device is increased for each
loop. The calculation of the average latency is adjusted.

Signed-off-by: Clara Kowalsky <clara.kowalsky@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>