xnthread_set_periodic: Improve period adjusting support
Currently, when a periodic task invokes xnthread_set_periodic() multiple
times, the timer is reset each time. Consequently, the period drifts
because the elapsed time is not taken into consideration.

To improve the situation, this patch alters the behavior when idate
is XN_INFINITE, to keep the current timeout value and alter the period
only after the current period expires.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>