testsuite: add latency treshold option to latency test
Add an -m option that allows to set a maximum latency treshhold to the
latency test.

If the option is used, the test will fail if the maximum measured
latency is larger than the treshold passed in microseconds.

This is mainly implemented for CI long term latency tests to
automatically let the pipeline fail if the measured latency is higher
than expected.

Signed-off-by: Tobias Schaffner <tobias.schaffner@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>