prepare-kernel.sh: Add reverse mode
There is currently no way to undo preparing of a kernel, except using
--outpatch and reverting the resulting patch later.
But in link mode the kernel tree needs manual cleaning,
which can be tedious if a git clean -f -d -x is not possible.

The new --reverse flag offers this feature, it removes all previously
created symlinks and directories.
To be able to undo effects by patch_append(), keep a .orig file link
as the script already does for symlinks in the linux tree.

Signed-off-by: Richard Weinberger <richard@nod.at>
[Jan: add missing init/Kconfig reversion, fix indention]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>