drivers/via-rhine: Migrate from rtdm_printk() and printk() to pr_*()
In this case only a single printk() call was missing the log level.
Migrating to keep in sync with other drivers.

The missing log level is addressed by migrating to the pr_*() API.

rtdm_printk() itself is just an alias for printk(). Migrating helps to
remove this function once all drivers were updated.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>