drivers/experimental/3c59x: Remove unused / broken driver
The driver was imported nearly 10 years ago but was actually unusable.
There was a missmatch in the kconfig option name. The Makefile was
using
  CONFIG_RTNET_DRV_3C59X while the Kconfig introduced
  CONFIG_XENO_DRIVERS_NET_DRV_3C59X

See
commit 106ffba7b55d ("rtnet: import")

The driver appeared on the radar while cleaning up printk() log levels.

The driver would require adjustments to the recent PCI subsystem API
to compile successfully. Instead of doing those adjustments the driver
is being removed. Seems nobody is using it anymore.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
[Jan: also clean gitlab-ci]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>