cobalt: Move relax related lostage_wakeup work from stack to thread
We cannot rely on the lostage_wakeup work being scheduled only
synchronously over the CPU of the relaxing thread. It thread might have
a CPU migration and other wakeup events pending and resume on different
CPU. As the lostage_wakeup work is currently on the stack, it can become
corrupt when xnthread_relax leaves earlier on the new CPU than the work
was processed on the old one - ugly crashes ahead as seen in [1].

Analogously to the signal wakeups, move the relax work into the xnthread
structure to avoid such corruptions. We intentionally refrain from
using irq_work_sync in the relaxation exit path to avoid starting to
wait a CPU that is already busy again with a lengthy RT task.

[1] https://lore.kernel.org/xenomai/770b1e4bc42cfa1ba5f4f37cc854956f3e78bb60.camel@siemens.com/T/#t

Backport of 95ccd795ec5db6c2d4d867b0f29fae964cffcf82, only affecting
dovetail kernels, I-pipe copies the work message.

Reported-by: JerryJun <jerryjun123@163.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>