testsuite/smokey/posix-mutex: Fix race-condition
In the testcase pi_nesting_contend_multicore the thread low may be unblocked to
early preventing the thread medium from running until the end of the testcase,
making the testcase fail, as the thread medium is expected to run before the
thread low exits.

This is fixed by ensuring that all threads are properly setup within the
init-phase of the testcase, by additional instances of smokey_barrier used to
signal that a particular thread has been started during the init-phase of the
testcase.

Signed-off-by: Fabian Scheler <Fabian.Scheler@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>