y2038: lib/cobalt: Fix link failure observed on Debian 13
The following build failure was detected on Debian sid/trixie:

/usr/bin/ld: .libs/smokey.tmp: in function `run_fpu_stress':
testsuite/smokey/fpu-stress/fpu-stress.c:84:(.text+0x51ed): undefined reference to `__real___clock_nanosleep_time64'
/usr/bin/ld: .libs/smokey.tmp: in function `run_sigdebug':
testsuite/smokey/sigdebug/sigdebug.c:281:(.text+0x1bd9d): undefined reference to `__real___nanosleep64'

The problem here is once again the changed behavior of the
__USE_TIME_BITS64 macro. Up to glibc 2.39 this was only set/defined when
building a compat application. So when time_t was "upgraded" from 4 to 8
bytes.

From glibc 2.40 onwards this macro will always be defined when time_t
is 8 bytes. So we basically run into the different code paths when
building 64 bit applications.

The reason for the linkage error is that __clock_nanosleep_time64() and
__nanosleep64() are only provided by glibc when building compat
applications.

We only have the additional time64_t wrappers for glibc in place,
so we limit the additional symbol handling to glibc as well. That allows
us to use glibc specific makros like __TIMESIZE to detect the compat
case.

The glibc commit that changed the behavior can be found at [1].

[1] https://sourceware.org/git/?p=glibc.git;a=commitdiff;h=a4ed0471d71739928a0d0fa3258b3ff3b158e9b9;hp=a0698a5e92ceeed3409d28623b1d599da6bc887d

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>