lib/cobalt/internal: Fix lock-ups of cobalt_monitor_wait over signals
This is a very long pending issue, actually since day #1 of this code.
We were checking the wrong return code for interruptions by Linux
signals. Rather than using the opret code that the cobalt_monitor_wait
syscalls set when the xnsynch_sleep_on was interrupted, we checked the
syscall return code that might have been set to -EINTR as well in case
prepare_for_signal found a pending signal while returning to userspace.
However, in that case the monitor was still entered, and our attempt to
enter it again from userspace caused a deadlock.

This was first reported in [1] over alchemy's rt_buffer_read, but it
could affect many more direct or indirect users of cobalt_monitor_wait
& Co.

[1] https://lore.kernel.org/xenomai/9d5bb262-0646-451e-a5af-82c0473a3003@yahoo.com/

Reported-by: Mauro Salvini <mau.salvi@yahoo.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>