cobalt: Add missing evaluations of __xn_get/put_user return codes
We should always give userspace feedback on the validity of a provided
pointer, rather than silently swallowing errors. Except for
cobalt_map_user, this includes adding an access_ok check as well.

As a side-effect, this resolves a compiler issue caused by [1] in
__cobalt_monitor_wait: At least gcc 12 on arm64 manages to skip an
important instruction during fixup of _[_xn]_put_user that causes a
register corruption down the road [2]. That is triggered by us ignoring
the return value, thus making the compiler drop the asm-goto block for
it which would have contained that critical instruction as well (as
duplicate who knows why). While this may very well be a compiler bug,
the triggering pattern of us remains invalid.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=7fd298d4b39d8d5fe99d56811a7ed78c7a5377d5
[2] https://lore.kernel.org/xenomai/87frltuz6j.fsf@xenomai.org

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>