cobalt/vfile: Plug race while opening
The procfs does not ensure atomicity of the open handlers. Therefore, it
is possible that two instances enter successfully because we check for
concurrency early but only mark a vfile as used at the end of the
handler. When we run into that, the refcnt may underflow on cleanup:

 kernel BUG at ../kernel/xenomai/vfile.c:555!

Protect that path via a regular Linux mutex.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>