testsuite/smokey/y2038: Plug race between interruption signaler and waiter
While testing pselect interruption, force signaler and waiter on the
same CPU and prioritize the latter over the former. This avoids a race
between sending the signal right after being unblocked from the
semaphore and blocking in pselect. Seen rather frequently now with
qemu-arm 4-core machines.

With the schedule-based ordering in place, we can safely drop the
nanosleep before signalling.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>