kernel/cobalt/posix: Stop supporting timeout parameter in recvmmsg
It turned out that trying to put a timer-based timeout on top of the
loop that calls recvmsg_rt multiple times is broken in an almost
unfixable way.

First of all, timing out by setting XNTIMEO and resuming the waiter
(=caller of recvmmsg) races with entering and leaving the blocking
state inside the respective driver. When the timer fires while the
thread is not inside xnthread_suspend, the condition is lost, and we may
even enter infinite blocking later on. This was observed in qemu setups
with overloaded hosts, but it could also happen in proper deployment if
the caller of recvmmsg is interrupted before suspending.

Second, the driver has no idea about the outer timeout and will
therefore happily use its own per-socket timeout in addition when set.
That can lead to unexpected results for the caller. Unfortunately, there
is no standard way in RTDM to control the socket timeout, not to speak
of an efficient one.

Finally, it would need an interface change in RTDM to collaborate with
drivers for implementing the out timeout without races and without
interference by the socket timeout. This appears not worth the effort.

Given that the socket timeout can still be used in place of the recvmmsg
timeout, though not with identical semantics, we avoid complex
refactoring in favor of discontinuing the timeout feature. Programs that
continue to pass non-zero timeout values will receive an error from now
on.

As this obsoletes some functional testing of the timeout via recvmmsg64,
adjust it.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>