lib/cobalt: Unbreak sigshadow installation for statically linked applications
When statically linking, we do not have a .eh_frame_hdr section. If we
then call backtrace before frame_dummy was invoked during late
constructions, we abort. Address that by moving the backtrace warm-up
call into a separate constructor that is armed earlier by
install_sigshadow, or call it later directly if install_sigshadow is
running after the construction phase.

Reported-by: Haiyang He <aspirestro@163.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>